requirements.txt
click==6.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.3
djangorestframework==3.7.7
factory-boy==2.10.0
Faker==0.8.12
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2
python-dateutil==2.7.2
python-decouple==3.1
pytz==2018.3
six==1.11.0
static3==0.7.0
text-unidecode==1.2
Werkzeug==0.14.1
whitenoise==3.3.1
gunicorn
